Model updates

+++ src/evkx.models/Models/CloudMedia.cs
@@ -106,7 +106,7 @@ namespace evdb.Models
         {
             if (Height.HasValue && Width.HasValue)
             {
-                return (int)((double)(400 / Width.Value) * (double)Height.Value);
+                return (int)((double)(400 / (double)Width.Value) * (double)Height.Value);
             }
             return default;
         }
